Rename project to `thermal_printer_plus` and update plugin name to `thermal_printer_plus_plugin`. On Windows
