+++ windows/include/thermal_printer_plus/thermal_printer_plus_plugin.h
